requirements.txt
sanic==19.9.0
click==7.0
aiohttp==3.6.2
sanic_oauth==0.4.0
sanic_mako==0.5.0
WTForms==2.2.1
Sanic-wtf==0.5.0
sanic-session==0.5.6
aioredis==1.3.1
cython==0.29.13
mistune==0.8.4
asyncblink==0.3.2
werkzeug==0.16.0
Pygments==2.4.2
cssmin==0.2.0
PyYAML==5.1.2
gunicorn==19.9.0
sanic-sentry==0.1.7
sanic-jwt==1.2.1
pangu==4.0.6.1
redis==3.3.11
PyYAML==5.1.2
aiosmtplib==1.1.0
arq==0.18.3
tortoise-orm==0.15.4
lxml==4.4.2
extraction==0.3
-e git+https://github.com/dongweiming/aiomysql.git@python38#egg=aiomysql
dj-database-url==0.5.0
Django==3.0.5
